Refactored MoreLikeThis component to a single folder
